refactor: 출발지점과 도착지점 사이에 기물이 막혀있는지 확인해주는 객체 생성
