+++ src/main/java/chess/domain/ChessBoard.java
@@ -4,11 +4,13 @@ import chess.domain.piece.Empty;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 public class ChessBoard {
 
-    Map<Position, Piece> board;
+    private final Map<Position, Piece> board;
 
     ChessBoard(Map<Position, Piece> board) {
         this.board = board;
@@ -21,6 +23,14 @@ public class ChessBoard {
 
     public Piece choicePiece(Position position) {
         return board.get(position);
+    }
+
+    public List<Piece> choiceBetweenPiece(List<Position> betweenPosition) {
+        List<Piece> betweenPiece = new ArrayList<>();
+        for (Position position : betweenPosition) {
+            betweenPiece.add(board.get(position));
+        }
+        return betweenPiece;
     }
 
     public Map<Position, Piece> getBoard() {
