+++ src/test/java/chess/domain/ChessGameTest.java
@@ -4,6 +4,7 @@ import chess.domain.piece.Camp;
 import chess.domain.piece.Empty;
 import chess.domain.piece.Pawn;
 import chess.domain.position.File;
+import chess.domain.position.Path;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
 import org.junit.jupiter.api.DisplayName;
@@ -24,7 +25,7 @@ class ChessGameTest {
         Position whiteRookPosition = Position.of(File.A, Rank.ONE);
         Position toPosition = Position.of(File.A, Rank.THREE);
 
-        assertThatThrownBy(() -> chessGame.move(whiteRookPosition, toPosition))
+        assertThatThrownBy(() -> chessGame.move(whiteRookPosition, toPosition, new Path()))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -37,7 +38,7 @@ class ChessGameTest {
         Position whiteKnightPosition = Position.of(File.B, Rank.ONE);
         Position toPosition = Position.of(File.A, Rank.THREE);
 
-        assertDoesNotThrow(() -> chessGame.move(whiteKnightPosition, toPosition));
+        assertDoesNotThrow(() -> chessGame.move(whiteKnightPosition, toPosition, new Path()));
     }
 
     @Test
@@ -49,7 +50,7 @@ class ChessGameTest {
         Position emptyPosition = Position.of(File.A, Rank.THREE);
         Position toPosition = Position.of(File.A, Rank.FOUR);
 
-        assertThatThrownBy(() -> chessGame.move(emptyPosition, toPosition))
+        assertThatThrownBy(() -> chessGame.move(emptyPosition, toPosition, new Path()))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -62,7 +63,7 @@ class ChessGameTest {
         ChessBoard chessBoard = ChessBoardMaker.create();
         ChessGame chessGame = new ChessGame(chessBoard, Camp.WHITE);
 
-        chessGame.move(whitePawnPosition, emptyPosition);
+        chessGame.move(whitePawnPosition, emptyPosition, new Path());
 
         assertThat(chessBoard.getBoard().get(whitePawnPosition)).isInstanceOf(Empty.class);
         assertThat(chessBoard.getBoard().get(emptyPosition)).isInstanceOf(Pawn.class);
